#12 reading latest state added

+++ src/OS.Smog.Domain/Sensors/Interpreter/Expressions/ConcentrationValidationExpression.cs
@@ -4,11 +4,14 @@ using OS.Smog.Dto.Sensors;
 
 namespace OS.Smog.Domain.Sensors.Interpreter.Expressions
 {
-    public abstract class ConcentrationValidationExpression : ValueRangeValidationExpression<float>, IExpression<PayloadInterpretationContext>
+    public abstract class ConcentrationValidationExpression : ValueRangeValidationExpression<double>,
+        IExpression<PayloadInterpretationContext>
     {
-        private const string ConcentrationError = "Concentration must be greater or equal to 0.0 and less or equal than 100.0";
+        private const string ConcentrationError =
+            "Concentration must be greater or equal to 0.0 and less or equal than 100.0";
+
         public abstract string Name { get; }
-        public abstract Func<Data, float?> ValueProvider { get; }
+        public abstract Func<Data, double?> ValueProvider { get; }
 
         public abstract string SIUnit { get; }
 
@@ -21,7 +24,7 @@ namespace OS.Smog.Domain.Sensors.Interpreter.Expressions
 
                 if (!value.HasValue) continue;
 
-                if (!ValueIsInRange(value.Value, 0.0f, float.MaxValue))
+                if (!ValueIsInRange(value.Value, 0.00, double.MaxValue))
                 {
                     context.Errors.Add($"{item.Timestamp} : {Name} {ConcentrationError} ({value.Value}[{SIUnit}])");
                 }
